Remove duplicated optimization files
Remove duplicated files for dead trees elimination and local
dead store elimination.
